==> 16S rRNA amplicon sequencing/4. Core microbiota.Rmd <==
---
title: "Core microbiota"
author: "Hendra"
---


library(microbiome)
library(dplyr)
library(phyloseq)
library(jeevanuDB)
----------------------------------------------------------------------------------------

# Prevalence of taxonomic groups
ps = base::readRDS("D:/S3-KU/Rice seed analysis/Phyloseq/Phy16s-riceseed.rds")

head(prevalence(ps, detection = 1/100, sort = TRUE))
head(prevalence(ps, detection = 1/100, sort = TRUE, count = TRUE))

# Core microbiota analysis
core.taxa.standard <- core_members(ps, detection = 0, prevalence = 50/100)
pseq.core <- core(ps, detection = 0, prevalence = .7)
pseq.core2 <- aggregate_rare(ps, "Genus", detection = 0, prevalence = 1)
core.taxa <- taxa(pseq.core)

# Core abundance and diversity
core.abundance <- sample_sums(core(ps, detection = .01, prevalence = .95))

transform <- microbiome::transform

rh <- subset_samples(ps, Source == "Origin") 
print(rh)

rh <- prune_taxa(taxa_sums(rh) > 0, rh)
print(rh)

rh.rel <- microbiome::transform(rh, "compositional")
print(rh.rel)


# This step change composition to 1:2

taxa_names(rh.rel)[1:2]
taxa_names(rh.rel) <- paste0("ASV", seq(ntaxa(rh.rel)))
taxa_names(rh.rel)[1:2]

core.taxa.standard <- core_members(rh.rel, detection = 0.01, prevalence = 90/100)
core.taxa.standard

rh.core <- core(rh.rel, detection = 0.01, prevalence = .9)
class(rh.core )

core.taxa <- taxa(rh.core)
class(core.taxa)

otu_table(rh.core)

otu_table(rh.core)["ASV286"]

tax_table(rh.core)

tax.mat <- tax_table(rh.core)
class(tax.mat)

tax.df <- as.data.frame(tax.mat)
class(tax.df)

tax.df$OTU <- rownames(tax.df)

core.taxa.class <- dplyr::filter(tax.df, rownames(tax.df) %in% core.taxa)
knitr::kable(head(core.taxa.class))


# Core line plots

det <- c(0, 0.1, 0.5, 2, 5, 20)/100
prevalences <- seq(.05, 1, .05)

plot_core(rh.rel, prevalences = prevalences, 
          detections = det, plot.type = "lineplot") + 
  xlab("Relative Abundance (%)") + 
  theme_bw()
---------------------------------------------------------------------------------------

# Core heatmaps

prevalences <- seq(.05, 1, .05)
detections <- round(10^seq(log10(1e-2), log10(.2), length = 10), 3)

gray <- gray(seq(0,1,length=5))

p1 <- plot_core(rh.rel,
  plot.type = "heatmap",
  colours = gray,
  prevalences = prevalences,
  detections = detections, min.prevalence = .9) +
  xlab("Detection Threshold (Relative Abundance (%)")

p1 <- p1 + theme_bw() + ylab("ASVs")
p1
----------------------------------------------------------------
# Using viridis color palette

library(viridis)
print(p1 + scale_fill_viridis())

==Genus level

library(RColorBrewer)
library(knitr)
df <- p1$data 

list <- df$Taxa 

tax <- as.data.frame(tax_table(rh.rel))
class(tax)

tax$ASV <- rownames(tax)

tax2 <- dplyr::filter(tax, rownames(tax) %in% list) 
head(tax2)

tax.unit <- tidyr::unite(tax2, Taxa_level,c("Kingdom", "Phylum", "Class", "Family", "Genus", "Species", "Confidence", "ASV"), sep = "_;", remove = TRUE)

tax.unit$Taxa_level <- gsub(pattern="[a-z]__",replacement="", tax.unit$Taxa_level)

df$Taxa <- tax.unit$Taxa_level
knitr::kable(head(df))

p1$data <- df

plot(p1 + theme(axis.text.y = element_text(face="italic")))

rh.rel.gen <- aggregate_taxa(rh.rel, "Genus")

any(taxa_names(rh.rel.gen) == "Unknown")

rh.rel.gen <- subset_taxa(rh.rel.gen, Genus!="Unknown")

library(RColorBrewer)
prevalences <- seq(.7, 1, .7)
detections <- round(10^seq(log10(1e-2), log10(.2), length = 10), 2)

p1 <- plot_core(rh.rel.gen, 
                plot.type = "heatmap", 
                colours = rev(brewer.pal(5, "RdBu")),
                prevalences = prevalences, 
                detections = detections, min.prevalence = .9) +
    xlab("Detection Threshold (Relative Abundance)")
p1 <- p1 + theme_bw() + ylab("ASVs")
p1

#Save
ggsave("Core mcrobiota new.tiff", width = 6, height = 4)
-------------------------------------------------------------------------------------------------------------------

taxa_names(rh.rel.gen) <- gsub("Bacteria_ Proteobacteria_ Gammaproteobacteria_", "", taxa_names(rh.rel.gen))
p2 <- plot_core(rh.rel.gen, 
                plot.type = "heatmap", 
                colours = rev(brewer.pal(6, "RdBu")),
                prevalences = prevalences, 
                detections = detections, min.prevalence = .9) +
    xlab("Detection Threshold (Relative Abundance)") + 
  theme_bw() +
  theme(axis.text.x = element_text(angle = 90, family = "Arial", size = 15, color = "black"),
        axis.text.y = element_text(family = "Arial", size = 15, face = "italic", color = "black", margin = margin(r = 3)),
        axis.title.x = element_text(family = "Arial", size = 15, face = "bold", color = "black"),
        axis.title.y = element_text(family = "Arial", size = 15, face = "bold", color = "black"),
        plot.title = element_text(family = "Arial", size = 15, face = "bold", color = "black"))
scale_y_discrete(expand = c(0.1, 0.1))  # Mengurangi ruang kosong di sumbu Y

p2

#save
ggsave("Core microbiota1.png", width = 10, height = 6, dpi = 600, bg = "transparent")
-------------------------------------------------------------------------------------------------------------------
