==> 16S rRNA amplicon sequencing/5. Diversity analysis.Rmd <==
---
title: "Bacterial diversity analysis"
author: "Hendra"
---


library(ggClusterNet)
library(phyloseq)
library(tidyverse)
library(dplyr)
library(tidyr)
library(rstatix)
library(ggpubr)
library(ggplot2)
-------------------------------------------------------------------------------------------------------------------

# 1. Alpha Diversity

psseed_origin = base::readRDS("D:/S3-KU/Rice seed analysis/Phyloseq/New Phylo/Phyasv-origin.rds")
sample_data(psseed_origin)

index = c("Shannon","Inv_Simpson","Pielou_evenness","Simpson_evenness" ,"Richness" ,"Chao1","ACE" )

ps = psseed_origin
samplesize = min(phyloseq::sample_sums(ps))
ps11  = phyloseq::rarefy_even_depth(ps,sample.size = samplesize)
mapping = phyloseq::sample_data(ps11)
ps11 = phyloseq::filter_taxa(ps11, function(x) sum(x ) >0 , TRUE); ps11

mapping$Allgroup = as.factor(mapping$Allgroup)
count = as.data.frame(t(ggClusterNet::vegan_otu(ps11)))
alpha=vegan::diversity(count, "shannon")
x = t(count)

Shannon = vegan::diversity(x)
Shannon
      
Inv_Simpson <- vegan::diversity(x, index = "invsimpson")
Inv_Simpson

S <- vegan::specnumber(x);S

S2 = rowSums(x>0)
Pielou_evenness <- Shannon/log(S)
Simpson_evenness <- Inv_Simpson/S
est <- vegan::estimateR(x)
Richness <- est[1, ]
Chao1 <- est[2, ]
ACE <- est[4, ]
report = cbind(Shannon, Inv_Simpson, Pielou_evenness, Simpson_evenness,
               Richness, Chao1,ACE)
head(report)

index = merge(mapping,report , by="row.names",all=F)
sel = c(match("Inv_Simpson",colnames(index)),
        match("Pielou_evenness",colnames(index)),
        match("Simpson_evenness",colnames(index)),
        match("Richness",colnames(index)),
        match("Chao1",colnames(index)),
        match("ACE",colnames(index)),
        match("Shannon",colnames(index)))


n = length(sel) + 3
data = cbind(data.frame(ID = 1:length(index$Allgroup),group = index$Allgroup),index[sel])
head(data)

result = EasyStat::MuiKwWlx2(data = data,num = c(3:(n -1)))
result1 = EasyStat::FacetMuiPlotresultBox(data = data,num = c(3:(n -1)),
                                          result = result,
                                          sig_show ="abc",
                                          ncol = 3 )
p1_1 = result1[[1]] +
  ggplot2::guides(fill = guide_legend(title = NULL))
p1_1
-------------------------------------------------------------------------------------------------------------------

#---------Another ggplot----#
alppath = paste(otupath,"/alpha/",sep = "")
dir.create(alppath)

group = "Group"
axis_order =  phyloseq::sample_data(ps)$Allgroup %>%unique();axis_order
#> [1] "Group1" "Group2" "Group3"

  samplesize = min(phyloseq::sample_sums(ps))
  if (samplesize == 0) {
    print("0 number sequence of some samples")
    print("median number were used")
    ps11  = phyloseq::rarefy_even_depth(ps,sample.size = samplesize)
  } else{
    ps11  = phyloseq::rarefy_even_depth(ps,sample.size = samplesize)
  }
  
  mapping = phyloseq::sample_data(ps11)
  ps11 = phyloseq::filter_taxa(ps11, function(x) sum(x ) >0 , TRUE); ps11

  # head(mapping)
  colnames(mapping) = gsub(group,"AA", colnames(mapping))
  
  mapping$Group = mapping$AA
  mapping$Group = as.factor(mapping$Allgroup)
  # mapping$Group 

  count = as.data.frame(t(ggClusterNet::vegan_otu(ps11)))

  alpha=vegan::diversity(count, "shannon")
  
  x = t(count) 
  # head(x)

  Shannon = vegan::diversity(x)  ##shannon
  Shannon

  Inv_Simpson <- vegan::diversity(x, index = "invsimpson")
  Inv_Simpson

  S <- vegan::specnumber(x);S  ##。S2 = rowSums(x>0)

  S2 = rowSums(x>0)
  
  # Pielou_evenness，Simpson_evenness
  Pielou_evenness <- Shannon/log(S)
  Simpson_evenness <- Inv_Simpson/S
  est <- vegan::estimateR(x)
  est <- vegan::estimateR(x)
  Richness <- est[1, ]
  Chao1 <- est[2, ]
  ACE <- est[4, ]
  report = cbind(Shannon, Inv_Simpson, Pielou_evenness, Simpson_evenness,
                 Richness, Chao1,ACE) 
  head(report)

  alp = merge(mapping,report , by="row.names",all=F)

index = c("Shannon","Inv_Simpson","Pielou_evenness","Simpson_evenness" ,"Richness" ,"Chao1","ACE" )
#--
index= alp
# head(index)

# theme definition

mytheme1 <- theme(
  axis.text = element_text(size = 12, color = "black"),
  axis.title = element_text(size = 14, face = "bold"),
  legend.position = "right",
  panel.background = element_rect(fill = "white"),
  panel.border = element_rect(color = "black", fill = NA)
)

sel = c(match("Shannon",colnames(index)),match("Inv_Simpson",colnames(index)),match("Chao1",colnames(index)),match("Pielou_evenness",colnames(index)),match("ACE",colnames(index)))
data = cbind(data.frame(ID = 1:length(index$Allgroup),group = index$Allgroup),index[sel])
# head(data)

result = EasyStat::MuiaovMcomper2(data = data,num = c(3:5))

result1 = EasyStat::FacetMuiPlotresultBox(data = data,num = c(3:5),
                                result = result,
                                sig_show ="abc",ncol = 3 )
p1_1 = result1[[1]] + 
  ggplot2::scale_x_discrete(limits = axis_order) + 
  mytheme1 +
  ggplot2::guides(fill = guide_legend(title = NULL)) +
  ggplot2::scale_fill_manual(values = colset3)
p1_1

p1_1 = result1[[1]] +
    ggplot2::scale_x_discrete(limits = axis_order) + 
  mytheme1 +
  guides(
    color = guide_legend(title = NULL),
    shape = guide_legend(title = NULL),
    fill = guide_legend(title = NULL)
  ) +
  ggplot2::guides(fill = guide_legend(title = NULL)) +
  theme(
    strip.text = element_text(family = "Arial", face = "bold", size = 15),
    axis.title.x = element_blank(),
    axis.text = element_text(family = "Arial",  face = "bold", size = 18),
    axis.title.y = element_text(family = "Arial", face = "bold", size = 12),
    legend.text = element_text(family = "Arial", size = 12),
    legend.title = element_blank(),  # Menghapus title legend
    legend.key = element_rect(fill = NA, colour = NA),
    panel.border = element_rect(colour = "black", fill = NA, size = 1)
  )

p1_1

#save
FileName <- paste(alppath,"Alpha-all sample3", ".png", sep = "")
ggsave(FileName, p1_1, width = 8, height =7,dpi = 800)
-------------------------------------------------------------------------------------------------------------------

p1_1 = result1[[2]] %>% ggplot(aes(x = group , y = dd )) + 
  geom_violin(alpha = 1, aes(fill = group)) +
  geom_jitter(aes(color = group), position = position_jitter(0.19), size = 4, alpha = 0.5) +
  labs(x = "", y = "") +
  facet_wrap(.~name, scales = "free_y", ncol = 3) +
  geom_text(aes(x = group , y = y , label = stat)) +
  ggplot2::scale_x_discrete(limits = axis_order) + 
  mytheme1 +
  guides(
    color = guide_legend(title = NULL),
    shape = guide_legend(title = NULL),
    fill = guide_legend(title = NULL)
  ) +
  ggplot2::guides(fill = guide_legend(title = NULL)) +
  theme(
    # Judul indeks (facet label) dengan font Arial, bold, ukuran 15
    strip.text = element_text(family = "Arial", face = "bold", size = 15),
    
    # Menghapus keterangan sumbu X
    axis.title.x = element_blank(),
    axis.text = element_text(family = "Arial",  face = "bold", size = 18),
    
    # Menyesuaikan sumbu Y menjadi hitam, Arial, bold, 15
    axis.title.y = element_text(family = "Arial", face = "bold", size = 12),
    
    # Menyesuaikan legend dengan font Arial, bold, 15
    legend.text = element_text(family = "Arial", size = 12),
    legend.title = element_blank(),  # Menghapus title legend
    legend.key = element_rect(fill = NA, colour = NA),
    panel.border = element_rect(colour = "black", fill = NA, size = 1)
  )

p1_1

#save
FileName <- paste(alppath,"Alpha-all sample2", ".png", sep = "")
ggsave(FileName, p1_1, width = 8, height =7,dpi = 800)


res = EasyStat::FacetMuiPlotresultBar(data = data,num = c(3:5),result = result,sig_show ="abc",ncol = 3)
p1_2 = res[[1]]+ scale_x_discrete(limits = axis_order) + guides(color = FALSE) +
  mytheme1+ 
  guides(fill = guide_legend(title = NULL))+
  scale_fill_manual(values = colset1)
p1_2


res = EasyStat::FacetMuiPlotReBoxBar(data = data,num = c(3:5),result = result,sig_show ="abc",ncol = 3)
p1_3 = res[[1]]+ scale_x_discrete(limits = axis_order) + 
  mytheme1 + 
  guides(fill = guide_legend(title = NULL))+
  scale_fill_manual(values = colset1)
p1_3

FileName <- paste(alppath,"Alpha_Facet_box", ".pdf", sep = "")
ggsave(FileName, p1_1, width = ((1 + gnum) * 3), height =4,limitsize = FALSE)

FileName <- paste(alppath,"Alpha_Facet_bar", ".pdf", sep = "")
ggsave(FileName, p1_2, width = ((1 + gnum) * 3), height = 4,limitsize = FALSE)

FileName <- paste(alppath,"Alpha_Facet_boxbar", ".pdf", sep = "")
ggsave(FileName, p1_3, width = ((1 + gnum) * 3), height = 4,limitsize = FALSE)

FileName <- paste(alppath,"Alpha_Facet_box", ".jpg", sep = "")
ggsave(FileName, p1_1, width = ((1 + gnum) * 3), height =4,limitsize = FALSE)

FileName <- paste(alppath,"Alpha_Facet_bar", ".jpg", sep = "")
ggsave(FileName, p1_2, width = ((1 + gnum) * 3), height = 4,limitsize = FALSE)

FileName <- paste(alppath,"Alpha_Facet_boxbar", ".jpg", sep = "")
ggsave(FileName, p1_3, width = ((1 + gnum) * 3), height = 4,limitsize = FALSE)
-------------------------------------------------------------------------------------------------------------------

krusk1 = ggpubr::compare_means( Shannon ~ group, data=data, method = "kruskal.test")
krusk2 = ggpubr::compare_means( Richness ~ group, data=data, method = "kruskal.test")
krusk3 = ggpubr::compare_means( Pielou_evenness ~ group, data=data, method = "kruskal.test")

dat = rbind(krusk1,krusk2,krusk3) %>% as.data.frame()
FileName <- paste(alppath,"/alpha_diversity_index_all_p_Kruskal-Wallis.csv", sep = "")
write.csv(dat,FileName,sep = "")
-------------------------------------------------------------------------------------------------------------------

# 2. Beta Diversity

library(ggClusterNet)
library(phyloseq)
library(tidyverse)
library(jeevanuDB)
-------------------------------------------------------------------------------------------------------------------

psseed_origin = base::readRDS("D:/S3-KU/Rice seed analysis/Phyloseq/Phy16s-riceseed-origin.rds")
sample_data(psseed_origin)
psorigin <- subset_samples(ps,  Source == "Origin")
betapath = paste(otupath,"/beta/",sep = "")
dir.create(betapath)

group = "Group"
dist = "bray"
method ="PCoA"
Micromet = "adonis"
pvalue.cutoff = 0.05
pair=TRUE

ps1_rela = phyloseq::transform_sample_counts(psorigin, function(x) x / sum(x) )
class(ps1_rela)
str(ps1_rela)

is_dist <- inherits(unif, "dist")
print(is_dist)

# An example of a commonly used distance method
dist <- "bray"  # atau "jaccard"
unif <- phyloseq::distance(ps1_rela, method = dist, type = "samples")

# Check whether the distance result is a valid object
# Check the class
class(unif)          # Should return "dist"

# Inspect the structure
str(unif)            # View internal structure

# Verify if it's a valid distance matrix
inherits(unif, "dist")   # TRUE if it's a distance object, FALSE otherwise

if (method == "DCA") {
  # method = "DCA"
  ordi = phyloseq::ordinate(ps1_rela, method=method, distance=dist)
  #
  points = ordi$rproj[,1:2]
  colnames(points) = c("x", "y") #命名行名
  #
  eig = ordi$evals^2
}

#---------CCA----
if (method == "CCA") {
  # method = "CCA"
  ordi = phyloseq::ordinate(ps1_rela, method=method, distance=dist)
  #
  points = ordi$CA$v[,1:2]
  colnames(points) = c("x", "y") #
  #
  eig = ordi$CA$eig^2
}

#---------RDA----
if (method == "RDA") {
  # method ="RDA"
  ordi = phyloseq::ordinate(ps1_rela, method=method, distance=dist)

  points = ordi$CA$v[,1:2]
  colnames(points) = c("x", "y") #
  
  eig = ordi$CA$eig
}

#---------MDS----
if (method == "MDS") {
  # method = "MDS"
  # ordi = ordinate(ps1_rela, method=ord_meths[i], distance=dist)
  ordi = phyloseq::ordinate(ps1_rela, method=method, distance=dist)
  points = ordi$vectors[,1:2]
  colnames(points) = c("x", "y") #
  eig = ordi$values[,1]
}

if (method == "PCoA") {
  # Calculate the distance matrix
  unif = phyloseq::distance(ps1_rela, method = dist, type = "samples")
  
  # Check if the distance matrix is valid
  if (!inherits(unif, "dist")) {
    stop("The distance matrix is not valid. Please check the input data or method.")
  }
  
  # Perform PCoA
  pcoa = stats::cmdscale(unif, k = 2, eig = TRUE)
  
  # Convert the results to a data frame
  points = as.data.frame(pcoa$points)
  colnames(points) = c("x", "y")
  
  # Save the eigenvalues
  eig = pcoa$eig
}

if (method == "PCoA") {
  # Calculate distance matrix
  unif <- phyloseq::distance(ps1_rela, method = dist, type = "samples")
  
  # Perform PCoA
  pcoa <- stats::cmdscale(unif, k = 2, eig = TRUE)
  
  # Extract coordinates
  points <- as.data.frame(pcoa$points)
  colnames(points) <- c("x", "y")
  
  # Extract eigenvalues
  eig <- pcoa$eig
}

str(x)  # Check structure of x
str(y)  # Check structure of y

#----PCA----

otu_table = as.data.frame(t(ggClusterNet::vegan_otu(ps1_rela )))
# head(otu_table)
# method = "PCA"
if (method == "PCA") {
  otu.pca = stats::prcomp(t(otu_table), scale.default = TRUE)
  #
  points = otu.pca$x[,1:2]
  colnames(points) = c("x", "y") #
  # 
  # otu.pca$rotation
  # 
  eig=otu.pca$sdev
  eig=eig*eig
}

# method = "LDA"
#---------------LDA----
if (method == "LDA") {
  #
  # library(MASS)
  data = t(otu_table)
  # head(data)
  data = as.data.frame(data)
  # data$ID = row.names(data)
  data = scale(data, center = TRUE, scale = TRUE)
  dim(data)
  data1 = data[,1:10]
  map = as.data.frame(sample_data(ps1_rela))
  model = MASS::lda(data, map$Group)

  ord_in = model
  axes = c(1:2)
  points = data.frame(predict(ord_in)$x[, axes])
  colnames(points) = c("x", "y") #
  #-----
  eig= ord_in$svd^2
}

#---------------NMDS----
# method = "NMDS"

if (method == "NMDS") {
  #---------
  # i = 5
  # dist = "bray"
  ordi = phyloseq::ordinate(ps1_rela, method=method, distance=dist)
  points = ordi$points[,1:2]
  colnames(points) = c("x", "y") #
  #stress
  stress = ordi$stress
  stress= paste("stress",":",round(stress,2),sep = "")
}

#---------------t-sne----
# method = "t-sne"
if (method == "t-sne") {
  data = t(otu_table)
  # head(data)
  data = as.data.frame(data)
  # data$ID = row.names(data)
  #
  data = scale(data, center = TRUE, scale = TRUE)

  dim(data)
  map = as.data.frame(sample_data(ps1_rela))
  # row.names(map)
  #---------tsne
 
# install.packages("Rtsne")
  library(Rtsne)

  tsne = Rtsne::Rtsne(data,perplexity = 3)

  points = as.data.frame(tsne$Y)
  row.names(points) =  row.names(map)
  colnames(points) = c("x", "y") #命名行名
  stress= NULL
}

g = sample_data(psorigin)$Allgroup %>% unique() %>% length()
n = sample_data(psorigin)$Allgroup%>% length()
o = n/g

source("D:/S3-KU/Taowen analysis/MicroTest.R")
source("D:/S3-KU/Taowen analysis/pairMicroTest.R")

if (Micromet == "adonis") {
  # Ensure metadata matches distance matrix
  map <- map[rownames(map) %in% rownames(as.matrix(unif)), , drop = FALSE]
  
  # Run Adonis
  ado <- vegan::adonis2(unif ~ map$Allgroup, permutations = 999)
  
  R2 <- paste("Adonis:R ", round(ado$R2[1], 3), sep = "")
  p_v <- paste("p: ", ado$`Pr(>F)`[1], sep = "")
  title1 <- paste(R2, " ", p_v, sep = "")
  
  title1
}

if (F) {
  pairResult = pairMicroTest(ps = ps1_rela, Micromet = Micromet, dist = dist)
  
} else {
  pairResult = "no result"
}

map = as.data.frame(phyloseq::sample_data(ps1_rela))
map$Allgroup = as.factor(map$Allgroup)
colbar = length(levels(map$Allgroup))

points = cbind(points, map[match(rownames(points), rownames(map)), ])
# head(points)
points$ID = row.names(points)


if (method %in% c("DCA", "CCA", "RDA",  "MDS", "PCoA","PCA","LDA")) {
  p2 =ggplot(points, aes(x=x, y=y, fill = Allgroup)) +
    geom_point(alpha=.7, size=5, pch = 21) +
    labs(x=paste0(method," 1 (",format(100*eig[1]/sum(eig),digits=4),"%)"),
         y=paste0(method," 2 (",format(100*eig[2]/sum(eig),digits=4),"%)"),
         title=title1) +
    stat_ellipse(linetype=2,level=0.68,aes(group=Allgroup, colour=Allgroup))

  p3 = p2+ggrepel::geom_text_repel(aes(label=points$ID),size = 5)
  p3
}

if (method %in% c("DCA", "CCA", "RDA", "MDS", "PCoA", "PCA", "LDA")) {
  p2 = ggplot(points, aes(x = x, y = y, fill = Allgroup)) +
    geom_point(alpha = 1, size = 5, pch = 21) +
    labs(
      x = paste0(method, " 1 (", format(100 * eig[1] / sum(eig), digits = 4), "%)"),
      y = paste0(method, " 2 (", format(100 * eig[2] / sum(eig), digits = 4), "%)"),
      title = title1
    ) +
    stat_ellipse(linetype = 1.5, level = 0.65, aes(group = Allgroup, colour = Allgroup), linewidth = 1.3) +
  theme(
    panel.background = element_rect(fill = NA, colour = NA),  # Transparent panel
    plot.background = element_rect(fill = NA, colour = NA),   # Transparent plot background
    legend.position = c(1.87, 1.20),                          # Legend at bottom-right corner
    legend.key = element_rect(fill = NA, colour = NA),        # Transparent legend key
    legend.title = element_blank(),                           # Remove legend title
    legend.text = element_text(family = "Arial", size = 12),  # Arial, bold, size 12 for legend text
    panel.grid.major = element_line(colour = NA, size = 0.1), # Light major grid lines
    panel.grid.minor = element_line(colour = NA, size = 0.1), # Light minor grid lines
    axis.text = element_text(family = "Arial", face = "bold", size = 18),  # Arial, bold, size 18 for axis text
    axis.title.x = element_text(family = "Arial", face = "bold", size = 20), # Arial, bold for X axis title
    axis.title.y = element_text(family = "Arial", face = "bold", size = 20), # Arial, bold for Y axis title
    axis.line = element_line(colour = "black"),               # Black axis lines
    panel.border = element_rect(colour = "black", fill = NA, size = 1)  # Add border around panel with thickness 1
  )

p3 = p2 + ggrepel::geom_text_repel(aes(label = points$ID), family = "Arial", size = 4)
p3
}

#save
  FileName1 <- paste(betapath,"/a2_",method,"beta-sample2.png", sep = "")
  ggsave(FileName1 , p3, width = 7, height = 6, dpi = 800)
------------------------------------------------------------------------------------------------------------------------

if (method %in% c("DCA", "CCA", "RDA", "MDS", "PCoA", "PCA", "LDA")) {
  p2 = ggplot(points, aes(x = x, y = y, fill = Allgroup)) +
    geom_point(alpha = .7, size = 5, pch = 21) +
    labs(
      x = paste0(method, " 1 (", format(100 * eig[1] / sum(eig), digits = 4), "%)"),
      y = paste0(method, " 2 (", format(100 * eig[2] / sum(eig), digits = 4), "%)"),
      title = title1
    ) +
    stat_ellipse(linetype = 2, level = 0.68, aes(group = Allgroup, colour = Allgroup)) +
    theme(
      panel.background = element_rect(fill = "white", colour = "black"),  # 
      panel.grid.major = element_line(colour = "grey80"),                 # 
      panel.grid.minor = element_line(colour = "grey90"),                 # 
      plot.background = element_rect(fill = "white"),                    # 
      legend.key = element_rect(fill = "white", colour = "black"),       # 
      legend.background = element_rect(fill = "white", colour = "black") # 
    )

  p2
}
-------------------------------------------------------------------------------------------------------------------
mytheme1 = ggplot2::theme_bw() + ggplot2::theme(
  panel.background=  ggplot2::element_blank(),
  panel.grid=  ggplot2::element_blank(),
  legend.position="right",
  legend.title =  ggplot2::element_blank(),
  legend.background= ggplot2::element_blank(),
  legend.key= ggplot2::element_blank(),
  plot.title =  ggplot2::element_text(vjust = -8.5,hjust = 0.1),
  axis.title.y =  ggplot2::element_text(colour = "black"),
  axis.title.x = ggplot2::element_text(),
  axis.text =  ggplot2::element_text(),
  axis.text.x =  ggplot2::element_text(),
  axis.text.y =  ggplot2::element_text(),
  legend.text =  ggplot2::element_text()
)

mytheme2 = ggplot2::theme_bw() + ggplot2::theme(
  panel.background=  ggplot2::element_blank(),
  panel.grid=  ggplot2::element_blank(),
  legend.position="right",
  
  legend.title =  ggplot2::element_blank(),
  legend.background=  ggplot2::element_blank(),
  legend.key= ggplot2::element_blank(),
  plot.title =  ggplot2::element_text(vjust = -8.5,hjust = 0.1),
  axis.title.y =  ggplot2::element_text(),
  axis.title.x = ggplot2::element_text(),
  axis.text =  ggplot2::element_text(),
  axis.text.x =  ggplot2::element_text(angle = 90),
  axis.text.y =  ggplot2::element_text(),
  legend.text =  ggplot2::element_text()
)
-------------------------------------------------------------------------------------------------------------------
colset1 <- RColorBrewer::brewer.pal(9,"Set1")
colset2 <- RColorBrewer::brewer.pal(12,"Paired")
colset3 <- c(RColorBrewer::brewer.pal(11,"Set1"),RColorBrewer::brewer.pal(9,"Pastel1"))
colset4 = colset3
-------------------------------------------------------------------------------------------------------------------

if (method %in% c("NMDS")) {
  p2 =ggplot(points, aes(x=x, y=y, fill = Group)) +
    geom_point(alpha=.7, size=5, pch = 21) +
    labs(x=paste(method,"1", sep=""),
         y=paste(method,"2",sep=""),
         title=stress)+
    stat_ellipse( linetype = 2,level = 0.65,aes(group  =Group, colour =  Group))
  
  p3 = p2+ggrepel::geom_text_repel( aes(label=points$ID),size=4)
  p3
  if (method %in% c("t-sne")) {
    supp_lab = labs(x=paste(method,"1", sep=""),y=paste(method,"2",sep=""),title=title)
   p2 = p2 + supp_lab
   p3 = p3 + supp_lab
  }
  eig = NULL
}

if (method %in% c("t-sne")) {
  p2 =ggplot(points, aes(x=x, y=y, fill = Group)) +
    geom_point(alpha=.7, size=5, pch = 21) +
    labs(x=paste(method,"1", sep=""),
         y=paste(method,"2",sep=""))+
    stat_ellipse( linetype = 2,level = 0.65,aes(group  =Group, colour =  Group))
  
  p3 = p2+ggrepel::geom_text_repel( aes(label=points$ID),size=4)
  p3
  if (method %in% c("t-sne")) {
    supp_lab = labs(x=paste(method,"1", sep=""),y=paste(method,"2",sep=""),title=title1)
    p2 = p2 + supp_lab
    p3 = p3 + supp_lab
  }
  eig = NULL
}

  p3_1 = p2 + 
    scale_fill_manual(values = colset1)+
    scale_color_manual(values = colset1,guide = F) +
    mytheme1 + 
    theme(legend.position = c(0.8,0.15))
  p3_1
-------------------------------------------------------------------------------------------------------------------
  p3_2 = p3 +
    scale_fill_manual(values = colset1)+
    scale_color_manual(values = colset1,guide = F) + 
    mytheme1 + 
    theme(legend.position = c(0.8,0.15))
  p3_2

#save
  FileName1 <- paste(betapath,"/a2_",method,"beta-sample.png", sep = "")
  ggsave(FileName1 , p3_2, width = 10, height = 8, dpi = 600)
-------------------------------------------------------------------------------------------------------------------
